Consolidate logic around services
We move shared logic for services down to the Server class.
Additionally we remove special casing of the nanny in service_ports, and
instead tack on a nanny attribute to the Worker and WorkerState
directly.
