+++ distributed/node.py
@@ -1,7 +1,10 @@
 from __future__ import print_function, division, absolute_import
 
+import warnings
+
 from tornado.ioloop import IOLoop
 
+from .compatibility import unicode
 from .core import Server, ConnectionPool
 from .versions import get_versions
 
@@ -78,3 +81,52 @@ class ServerNode(Node, Server):
 
     def versions(self, comm=None, packages=None):
         return get_versions(packages=packages)
+
+    def start_services(self, default_listen_ip):
+        if default_listen_ip == "0.0.0.0":
+            default_listen_ip = ""  # for IPV6
+
+        for k, v in self.service_specs.items():
+            listen_ip = None
+            if isinstance(k, tuple):
+                k, port = k
+            else:
+                port = 0
+
+            if isinstance(port, (str, unicode)):
+                port = port.split(":")
+
+            if isinstance(port, (tuple, list)):
+                if len(port) == 2:
+                    listen_ip, port = (port[0], int(port[1]))
+                elif len(port) == 1:
+                    [listen_ip], port = port, 0
+                else:
+                    raise ValueError(port)
+
+            if isinstance(v, tuple):
+                v, kwargs = v
+            else:
+                kwargs = {}
+
+            try:
+                service = v(self, io_loop=self.loop, **kwargs)
+                service.listen(
+                    (listen_ip if listen_ip is not None else default_listen_ip, port)
+                )
+                self.services[k] = service
+            except Exception as e:
+                warnings.warn(
+                    "\nCould not launch service '%s' on port %s. " % (k, port)
+                    + "Got the following message:\n\n"
+                    + str(e),
+                    stacklevel=3,
+                )
+
+    def stop_services(self):
+        for service in self.services.values():
+            service.stop()
+
+    @property
+    def service_ports(self):
+        return {k: v.port for k, v in self.services.items()}
